+++ compaqt/main/validation.c
@@ -8,22 +8,26 @@
     if (b->offset + length > b->allocated) return 1; \
 } while (0)
 
-#define BUFFER_REFRESH() do { \
-    if (b->offset + MAX_METADATA_SIZE > b->allocated) \
+#define BUFFER_REFRESH(length) do { \
+    if (b->offset + length >= b->allocated) \
     { \
-        if (fseek(file, b->offset, SEEK_CUR) != 0) return 1; \
-        if ((b->allocated = fread(b->msg, 1, b->allocated, file)) == 0) return 1; \
+        b->allocated = fread(b->msg, 1, b->allocated, file); \
+        if (b->allocated == 0) return 1; /* EOF or error */ \
+        if (b->allocated <= length) return 1; /* File not big enough */ \
         b->offset = 0; \
     } \
-    else if (b->offset + 1 > b->allocated) return 1; \
+} while (0)
+
+#define CHECK(length) do { \
+    if (file != NULL) \
+        BUFFER_REFRESH(length); \
+    else \
+        OVERREAD_CHECK(length); \
 } while (0)
 
 static inline int _validate(buffer_t *b, FILE *file)
 {
-    if (file != NULL)
-        BUFFER_REFRESH();
-    else
-        OVERREAD_CHECK(1);
+    CHECK(0);
     
     const char dt_mask = RD_DTMASK(b);
     switch (dt_mask)
@@ -33,7 +37,7 @@ static inline int _validate(buffer_t *b, FILE *file)
         switch (RD_DTMASK_GROUP(b))
         {
         // Static lengths, no metadata to read, already incremented by `RD_DTMASK_GROUP`
-        case DT_FLOAT: OVERREAD_CHECK(8);  b->offset += 8;  return 0;
+        case DT_FLOAT: CHECK(8);  b->offset += 8; return 0;
         case DT_BOOLF: // Boolean and NoneType values don't have more bytes, nothing to increment
         case DT_BOOLT:
         case DT_NONTP: return 0;
@@ -45,30 +49,27 @@ static inline int _validate(buffer_t *b, FILE *file)
     {
         size_t num_items = 0;
         RD_METADATA(b->msg, b->offset, num_items);
-        OVERREAD_CHECK(0);
 
         // Multiply dict number of items by 2 as it has pairs
         if (dt_mask == DT_DICTN)
             num_items *= 2;
-
+        
         for (size_t i = 0; i < num_items; ++i)
-            if (_validate(b, file) == 1) return 1;
+            if (_validate(b, file) != 0) return 1;
         
         return 0;
     }
     case DT_EXTND:
     {
         ++(b->offset);
-        OVERREAD_CHECK(1);
+        CHECK(1);
 
         const size_t num_bytes = *(b->msg + b->offset++) & 0xFF;
-        OVERREAD_CHECK(num_bytes);
+        CHECK(num_bytes);
 
-        size_t length = 0;
-        memcpy(&length, b->msg + b->offset, num_bytes);
-        b->offset += num_bytes;
+        size_t length;
+        RD_METADATA_LM2(b->msg, b->offset, length, num_bytes);
 
-        OVERREAD_CHECK(length);
         b->offset += length;
 
         return 0;
@@ -76,11 +77,12 @@ static inline int _validate(buffer_t *b, FILE *file)
     case DT_NOUSE: return 1; // Not in use, so this would mean invalid
     default:
     {
-        // All other cases use VLE metadata, so read length and increment offset using it
-        size_t length = 0;
+        // All other cases use regular metadata, so read length and increment offset using it
+        size_t length;
         RD_METADATA(b->msg, b->offset, length);
-        OVERREAD_CHECK(length);
+
         b->offset += length;
+
         return 0;
     }
     }
@@ -124,7 +126,6 @@ PyObject *validate(PyObject *self, PyObject *args, PyObject *kwargs)
             return NULL;
         }
 
-        b.allocated = chunk_size;
         b.msg = (char *)malloc(chunk_size);
 
         if (b.msg == NULL)
@@ -134,7 +135,25 @@ PyObject *validate(PyObject *self, PyObject *args, PyObject *kwargs)
             return NULL;
         }
 
+        b.allocated = fread(b.msg, 1, chunk_size, file);
+
         result = _validate(&b, file);
+
+        // Do an extra overread check at the end
+        if (result == 0)
+        {
+            // The file offset before the last read
+            const size_t last_offset = ftell(file) - b.allocated;
+
+            // The file offset of the end of the file
+            fseek(file, 0, SEEK_END);
+            const size_t end_offset = ftell(file);
+
+            // Invalid data if we exceeded the highest possible offset (the end of the file)
+            if (last_offset + b.offset > end_offset)
+                result = 1;
+        }
+
         fclose(file);
         free(b.msg);
     }