+++ compaqt/py_fallback/__init__.py
@@ -1,0 +1,1 @@
+# Init file so Python can import files in this folder