+++ compaqt/py_fallback/exceptions.py
@@ -1,0 +1,7 @@
+# This file contains custom exceptions
+
+class EncodingError(Exception): pass
+class DecodingError(Exception): pass
+class ValidationError(Exception): pass
+class FileOffsetError(Exception): pass
+