+++ compaqt/py_fallback/settings.py
@@ -1,0 +1,5 @@
+from .unsupported import py_fallback_notimplemented
+
+manual_allocations = py_fallback_notimplemented
+dynamic_allocations = py_fallback_notimplemented
+