+++ compaqt/py_fallback/unsupported.py
@@ -1,0 +1,2 @@
+def py_fallback_notimplemented(*args, **kwargs):
+    raise NotImplementedError("This method is not supported in the Python fallback version of Compaqt")